+++ prisma/migrations/20260202175813_scores_duration/migration.sql
@@ -1,0 +1,2 @@
+-- AlterTable
+ALTER TABLE "scores" ADD COLUMN     "duration" INTEGER;